Overleaf saving implementation
I've implemented a script that will use selenium to log in to Overleaf,
download a specific project via the download link, save it to a backup
folder, unzip it, remove the zip file, and commit the changes to a git
repository. This is designed so in collaborative projects I can have git
versioning and backups for the project without having to pay overleaf
for their subscription. It's a smart workaround and I'm feeling a little
smug about it.

The credentials are managed in a 'creds.py' file which has been
gitignored, so feel free to use it yourself. You'll just need to create
the file with an object of your credentials.

The bash script is intended to run by crontab so this can happen if I
remember to do it or not.