Overleaf update
Updated the Overleaf readme and tried to make a tweak. Looks like
Overleaf was sending me captchas preventing the login from working, so
this approach won't cut it any more. Not sure if there are some
suspicious signals it would be possible to get around or if that's it
stuffed.